Recommendation for Authorization types change
In your change it looks like you're trying to us the file to pass
in all of the arguements and essentially do the equivalent of running
a bash script and foking a bunch of hey processes (one for each auth
header). I'd recommend reusing all of the same process internal of
the hey load generator and modifying the request before it gets set
to add one of the auth types. hey has a RequestFunc value in the Work
struct and that seems like the best way to override the default behavior.
Also, you are better off making the file being read in generic in case
you want to make the generated http requests more generic in the future.
For now it can just be a different auth type on each line, but in the
future you might make this a json file that allows you to change other
things like the body of the request or add other headers.